Фикс модульности
🔧 Что исправлено:
1. 🏗️ Создан core модуль routers/core/
✅ Содержит глобальные команды /start, /menu, main_menu
✅ Независим от других модулей - работает всегда
✅ Проверяет права доступа пользователей
✅ Показывает главное меню
2. 🔧 Исправлен id_module
❌ Было: @message() без фильтров - перехватывал ВСЕ
✅ Стало: @callback_query(F.data == "id_mode") - только свои команды
3. 🗑️ Очищен email модуль
Команды /start и /menu удалены
Инициализация состояния осталась при активации email режима
4. 📊 Обновлен порядок модулей
Apply to bot.py
🎯 Результат:
✅ Команды /start и /menu работают независимо от модулей
✅ Можно удалять любые модули без поломки бота
✅ Нет конфликтов между обработчиками
✅ Четкое разделение ответственности
📋 Теперь можно безопасно:
Удалять и добавлять модули без страха поломать основные функции
Команды /start и /menu будут работать всегда
Каждый модуль отвечает только за свою функциональность
Архитектура стала гораздо более устойчивой и модульной! 🎉
