+++ readme.md
@@ -114,6 +114,9 @@ project/
 │
 ├── routers/                # Все роутеры (модули логики)
 │   ├── __init__.py         # Автозагрузка роутеров
+│   ├── core/               # Ядро бота (команды /start, /menu)
+│   │   ├── __init__.py     # Экспорт core_router
+│   │   └── router.py       # Глобальные команды
 │   ├── email_router/       # Email функциональность (независимый модуль)
 │   │   ├── __init__.py     # Экспортирует router + MENU_CONFIG
 │   │   ├── router.py       # Логика роутера
@@ -209,6 +212,7 @@ routers/
 ## ⚙️ Возможности
 - Поддержка нескольких пользователей (`ALLOWED_USER_IDS`)
 - Отправка email с вложениями и текстом
+- **🏗️ Устойчивая архитектура** - глобальные команды в core модуле, независимо от других модулей
 - **🆔 Модуль получения ID** - показывает ID пользователя и ID чата/группы
 - **🧪 Тестовый модуль** - демонстрирует возможности системы (callback, text filters, emoji filters)
 - **🔧 Модульная архитектура** - true plug-and-play модули с автоматической регистрацией
