+++ routers/email_router/__init__.py
@@ -4,5 +4,5 @@ from .router import email_router
 MENU_CONFIG = {
     'text': '✉️ Отправка email',
     'callback_data': 'email_mode',
-    'order': 10
+    'order': 10  # После core модуля
 }
